+++ com.unity.shadergraph/Editor/VFX.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f93b01c64b6ff9f4abf87d9aff75f4bc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
