+++ com.unity.shadergraph/Editor/VFX/GraphUtilForVFX.cs.meta
