+++ com.unity.shadergraph/Editor/VFX/ShaderParser.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 10d6072648873bd4eb2c6e97279a44e3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
