+++ com.unity.shadergraph/Editor/VFX/ShaderParser/PassPart.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 522b926e11e68ed42974b9656242e9a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
