+++ com.unity.shadergraph/Editor/VFX/ShaderParser/ShaderDocument.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 27f68bdc992de41468567007b046a50e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
