+++ com.unity.shadergraph/Editor/VFX/ShaderParser/ShaderPart.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f576e254dae4d2a4ca9d0e2bc415b607
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
