+++ com.unity.shadergraph/Editor/VFX/ShaderParser/SubShaderPart.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 722914b38d1fd3a4895b3f441a704f53
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
